pyproject.toml
[tool.poetry]
name = "stock javis system"
version = "0.1.0"
description = ""
authors = ["leftnadal <136068537+leftnadal@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
python-dotenv = "^1.1.0"
django-cors-headers = "^4.9.0"
requests = "^2.32.5"
djangorestframework-simplejwt = "^5.5.1"
celery = "^5.5.3"
redis = "^7.0.1"
django-celery-beat = "^2.8.1"
django-celery-results = "^2.6.0"
channels = "^4.3.1"
channels-redis = "^4.3.0"
django-redis = "^6.0.0"
daphne = "^4.2.1"
pandas = "^2.3.3"
numpy = "^2.3.5"
psycopg2-binary = "^2.9.11"
neo4j = "^5.0.0"
anthropic = "^0.48.0"
sentence-transformers = "^5.2.0"
rank-bm25 = "^0.2.2"
google-genai = "^1.55.0"
httpx = "^0.28.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pytest = "^9.0.2"
pytest-django = "^4.11.1"
pytest-asyncio = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
